smallhr/requirements.txt
django
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
django-filter
django-environ
python-dotenv
django-redis
redis
celery
django-celery-beat
django-celery-results
gunicorn
whitenoise